deleted directories and added shared topology optimization modules to examples/topology_optimization/modules.py
