+++ examples/topology_optimization/ex_compliance_minimal.py
