+++ examples/topology_optimization/ex_compliance_prescribed_displacements.py
