+++ examples/topology_optimization/ex_compliant_mechanism.py
