+++ examples/topology_optimization/ex_volume_stress.py
@@ -14,7 +14,7 @@ DOI: https://doi.org/10.1007/s00158-016-1524-0
 import numpy as np
 
 import pymoto as pym
-from pymoto.modules.assembly import get_B, get_D
+from modules import Stress, VonMises, ConstraintAggregation
 
 # Problem settings
 nx, ny = 100, 50  # Domain size
@@ -26,86 +26,6 @@ maximum_vm_stress = 0.5
 displacement_constraint = False
 scaling_displacement_constraint = 10.0
 max_displacement = 100.0
-
-
-class ConstraintAggregation(pym.Module):
-    """
-    Unified aggregation and relaxation.
-
-    Implemented by @artofscience (s.koppen@tudelft.nl), based on:
-
-    Verbart, A., Langelaar, M., & Keulen, F. V. (2017).
-    A unified aggregation and relaxation approach for stress-constrained topology optimization.
-    Structural and Multidisciplinary Optimization, 55, 663-679.
-    DOI: https://doi.org/10.1007/s00158-016-1524-0
-    """
-
-    def _prepare(self, P=10):
-        self.P = P
-
-    def _response(self, x):
-        """
-        a = x + 1
-        b = aggregation(a)
-        c = b - 1
-        """
-        self.n = len(x)
-        self.x = x
-        self.y = self.x + 1
-        self.z = self.y ** self.P
-        z = ((1 / len(self.x)) * np.sum(self.z)) ** (1 / self.P)  # P-mean aggregation function
-        return z - 1
-
-    def _sensitivity(self, dfdc):
-        return (dfdc / self.n) * (np.sum(self.z) / self.n) ** (1 / self.P - 1) * self.y ** (self.P - 1)
-
-
-class VonMises(pym.Module):
-    def _prepare(self, *args, **kwargs):
-        # Vandermonde matrix
-        self.V = np.array([[1, -0.5, 0], [-0.5, 1, 0], [0, 0, 3]])
-
-    def _response(self, x):
-        self.x = x
-        self.y = (x.dot(self.V) * x).sum(1)
-        return np.sqrt(self.y)
-
-    def _sensitivity(self, dfdv):
-        return dfdv[:, np.newaxis] * (self.y ** (-0.5))[:, np.newaxis] * self.x.dot(self.V)
-
-
-class Stress(pym.Module):
-    def _prepare(self, E=1, nu=0.3, plane='strain', domain=pym.DomainDefinition, *args, **kwargs):
-        siz = domain.element_size
-        self.domain = domain
-
-        # Constitutive model
-        self.D = siz[2] * get_D(E, nu, plane.lower())
-
-        # Numerical integration
-        self.B = np.zeros((3, 8), dtype=float)
-        w = np.prod(siz[:domain.dim] / 2)
-        for n in domain.node_numbering:
-            pos = n * (siz / 2) / np.sqrt(3)  # Sampling point
-            dN_dx = domain.eval_shape_fun_der(pos)
-            self.B += w * get_B(dN_dx)
-
-        self.dofconn = domain.get_dofconnectivity(2)
-
-    def _response(self, u):
-        self.elemental_strain = self.B.dot(u[self.dofconn].transpose())
-        self.elemental_strain[2, :] *= 2  # voigt notation
-        return self.D.dot(self.elemental_strain).transpose()
-
-    def _sensitivity(self, dfdv):
-        dgdsstrainmat = np.einsum('jk,kl->jl', dfdv, self.D)
-        dgdsstrainmat[:, 2] *= 2
-        dgdue = np.einsum('ij,jl->il', dgdsstrainmat, self.B)
-
-        y = np.zeros(self.domain.nnodes * 2)
-        for i in range(0, self.domain.nel):
-            y[self.dofconn[i, :]] += dgdue[i, :]
-        return y
 
 
 if __name__ == "__main__":
